templates, include: explicit Isabelle semantics for connectors
To verify capDL refinement, we need to know the intended semantics for
CAmkES connectors. We use a convention of declaring such semantics in
a new attribute, `isabelle_connector_spec`, which should contain
an Isabelle term of type `connector`. This text is inserted verbatim
when emitting the `Arch_Spec` theory for a CAmkES assembly.

We also add semantics for most of the built-in connectors. These are
merely links to the same built-in connectors of CamkesAdlSpec.
However, other connectors can now declare custom semantics if needed.