parser: Enable the passing of a list to a query
Extended support to pass a list of dictionaries to a camkes query.
This enables the user to make multiple queries within the single
call. In turn the call result returns a list of the resolved
queries. This commit updates the query grammar and the usages of
the query object to support multiple queries.