Updated libsel4camkes to the new libvirtqueue
This modules now wraps the new libvirtqueue in projects_libs. The new
libvirtqueue provides scatter gather buffers as well as multi entry
rings, so this camkes lib aims at providing usable wrappers to do all
the plumbing, such as iterating through the scattered buffers,
scattering buffers into chunks, etc.

Since both users of the virtqueue are not necessarily in the same
address space, they need to share buffers as offset into the array of
shared memory. This wrapper also abstracts this as much as possible by
doing the conversion inside the wrapper functions.

So as a summary, this modules provides :

Conversion between pointer and offset in shared memory.

Taking a single buffer, splitting it into chunks and scattering it in
the virtqueue as a single ring entry. (Ex: a driver scatters a big chunk
of empty memory to be filled or read by the device).

Taking a handle (a ring entry) and iterating through it to gather back a
scattered buffer. (Ex: a driver gathers back some memory that was filled
by the device, or a device reads the data sent by the driver to perform
some operation).

Taking a handle (a ring entry) and a buffer, and iterating through the
scatter list, copying the content of the buffer into the scattered
arrays, and enqueuing it back into the used ring. (Ex: a device copies
some memory it has just read from the hardware into the buffers that
were passed for filling by the driver).