templates: 'Fix' next_page_multiple
Previously `next_page_multiple` calculated the smallest page multiple
that can fit a dataport by multiplying by 2 everytime from a base of
4096. This was because of the dataport size `static_asserts` were
checking if the size was aligned to a power of 2. Now that this isn't
the case anymore (commit 6c191556), we 'fix' `next_page_multiple` so
that it adds increments of 4096. This also helps in saving memory by
only allocating what we need.