camkes-gen: support CMakeForegroundComplexCommands
We previously set USES_TERMINAL during commands that render templates so
that a python debugger could be used if necessary. The cost of
USES_TERMINAL is that the jobs can only run one at a time. Using
CMakeForegroundComplexCommands to switch between foreground and
background allows this behavior to be more easily controlled.