templates,marshal: Remove usages of c_symbol()
c_symbol() generates a unique C symbol name that depends on a global
counter. This leads to confusing variable names that change frequently
based on any other calls to c_symbol in a project build process. In
cases where a globally unique name is not needed, using a local name is
better.