Add seL4DMASharedData connector
This connector sets up a dataport connector that is added to the DMA
pool that the camkes runtime tracks for each component. Only one end of
the connection is expected to control allocating and freeing the memory,
the other ends can only perform cache operations and get the physical
addresses.