Add global_rpc_endpoint_badges macro
This macro assigns badges for different connectors that share the
global-rpc-endpoint object for a component instance. This function could
be thought of as an iterator that returns a set of badges for each
connection.  If called for every connection that an instance belongs to,
then the total set of badges returned will be distinct. This allows many
different interfaces to be multiplexed over a single endpoint object.