seL4DTBHardware: Use global interface name
The MMIO frame objects that this connector creates need to have unique
names based on a connector or instance.interface name. str(me) will
return the full instance.interface name of the interface. Otherwise
interfaces with the same name in different component instances will
cause the underlying allocator to throw an error.