parser: Add GPIO query engine
This new query engine handles the new 'gpio' CAmkES queries that the new
seL4GPIOServer connector exposes for its access control mechanisms.

The engine takes in a YAML file containing a list of GPIO pins and sorts
out the 'gpio' queries so that the connector templates for the
seL4GPIOServer connector can generate the appropriate structures and
functions.