runner,templates: Fix generate_seL4_SignalRecv
This function was fixed before to resolve the issue where
seL4_SignalRecv did not exist as a seL4 kernel syscall. The result of
that fix was to generate separate seL4_Signal and seL4_Recv calls. This
altered the macro so that the macro also assigns the result of seL4_Recv
to a variable. However, the macro assumes that the variable that
contains the message info to be sent is the same as the variable that
will store the result of the seL4_Recv/seL4NBSendRecv call.

This isn't a good assumption and so this commit adds in an extra
parameter to the macro for the result variable and also fixes the
missing result assignment for the realtime case. Also update the macro
usage in rpc-connector.c.