libsel4camkes: Cache DMA frame paddr
The 'dma_pin' function will request for a frame's physical address via a
syscall each time it is called. This is a fine for a few calls over the
lifetime of a program, but beyond that is pretty costly. Thus, this is
an optimisation to cache the result of the call so that we can avoid
syscalls each time to need to 'pin' (the mappings don't get changed at
all) the piece of DMA memory.