camkes: support DTBs with one interrupt value
Seems on ARM this was always used with device trees that have 3 values
per interrupt or extended interrupts that have just one interrupt. Fix
the handling for extended interrupts to allow more than one. Add support
for device trees (e.g. RISC-V HiFive) that use only one value per
interrupt. This is a hack that add some smart guessing to accept them
also. More documentation is also added about the problems. The long term
fix should be properly parsing the device tree, as it is done in the
kernel already.

Signed-off-by: Axel Heider <axelheider@gmx.de>