camkes-gen.cmake: best-effort build step can error
When linking multiple components together to create a "grouped"
component we run some objcopy commands to remove unsupported and unused
sections such as the .eh_frame.  Some toolchains don't allow this and
cause objcopy to error but this shouldn't cause the build step to fail.
The build rule was already trying to catch the error according to its
preceding comment, but now it does this correctly.

Signed-off-by: Kent McLeod <kent@kry10.com>