parser: fix attribute_reference regex
The lexer showed syntax errors when working with attribute references
using '<-'. The regular expression for angle_string is being checked by
the lexer when encountering the '<' first.
So if the file to be parsed contains a '>' following the attribute
reference, the lexer will tokenize everything from '<-' to this '>',
leading to a syntax error.
Examples for this behaviour would be an import statement with angled
brackets or an export statement containing '->'.
The proposed fix adds further characters to be excluded in the regular
expression, to avoid the angle_string rule to be active in such cases.

Signed-off-by: Thomas Wittal <thomas.wittal@hensoldt-cyber.de>