component.template.h: provide X_get_sender_id
Declare the function get_sender_id for all functions the component
provides. The function is only implemented in RPC connections. It will
lead to a link-time error when used for other connections.

We need to declare the function to make it available to component code
(-Wimplicit-function-declaration is now an error). Ideally the RPC
connector would take care of that, but it does not generate its own .h
file.

Signed-off-by: Gerwin Klein <gerwin.klein@proofcraft.systems>